Card Class 생성
