+++ PokerApp/Card_Generate_Test/Card_Generate_Test.swift
@@ -1,0 +1,37 @@
+//
+//  Card_Generate_Test.swift
+//  Card_Generate_Test
+//
+//  Created by 이준형 on 2021/01/06.
+//  Copyright © 2021 이준형. All rights reserved.
+//
+
+import XCTest
+
+class Card_Generate_Test: XCTestCase {
+
+    override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+
+        let card1 = Card(shape: .heart, number: .six)
+    }
+
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+
+        print(card1.description)
+    }
+
+    func testPerformanceExample() throws {
+        // This is an example of a performance test case.
+        measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
