+++ PokerApp/PokerApp.xcodeproj/xcuserdata/harrison.tom.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,6 +4,11 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
+		<key>Card_Generate_Test.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>1</integer>
+		</dict>
 		<key>PokerApp.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
