+++ README.md
@@ -1,3 +1,2 @@
 # swift-w1-poker
-# made by harrison
 모바일 1주차 카드게임 저장소
